Remove pencil weirdness in gitcommit textwidth
I thought pencil would correct set the textwidth in a git commit message
to 72 characters, but it seems not.

For some reason the pencil plugin was resetting it to tw=0, which is
wrong.

I'm explicitly setting textwidth=72 in the ftplugin for now, and
removing pencil init.

Just let tpope's git plugin do its thing.